+++ src/modules/buyflow/components/IdentityStep/IdentityStep.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { Field } from '../../../form-controls/components/Field/Field';
+import { Field } from '../../../form-controls/components';
 import { NavigationButtons } from '../NavigationButtons';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
